Fixing permission error

+++ backend/src/Services/projectService.js
@@ -53,39 +53,50 @@ const createProject=async(title, startDate, dueDate, assignedUserId)=>{
         throw new Error(`check error ${error}`)
     }
 }
-const updateProject=async(id,title, startDate, dueDate, assignedUserId)=>{
+const updateProject=async(userId, id,title, startDate, dueDate, assignedUserId)=>{
     try{
-        const result=await db.Project.update(
-            {
-                title:title,
-                createdAt:startDate,
-                endedAt:dueDate
-            },
-            {
-                where:{
-                  id:id
-                }
-            }
-        )
-        await db.ProjectMember.destroy({
-            where: {
-                projectId: id,
-                userId: { [Op.notIn]: assignedUserId} 
+        const checkRole=await db.ProjectMember.findOne({
+            where:{
+                userId:userId,
+                projectId:id
             }
         })
-        const existing = await db.ProjectMember.findAll({
-                where: { projectId: id }
-        });
-        const existingMemberIds = existing.map(row => row.userId);
-        console.log(existingMemberIds)
-        const newMembers = assignedUserId.filter(id => !existingMemberIds.includes(id));
-        console.log(newMembers)
-         await db.ProjectMember.bulkCreate(
-            newMembers.map(user_id => ({
-                projectId: id,
-                userId: user_id,
-            }))
-        );
+        const plainResult=await checkRole.get({plain:true})
+        if(plainResult.role==='admin'){
+            const result=await db.Project.update(
+                {
+                    title:title,
+                    createdAt:startDate,
+                    endedAt:dueDate
+                },
+                {
+                    where:{
+                    id:id
+                    }
+                }
+            )
+            await db.ProjectMember.destroy({
+                where: {
+                    projectId: id,
+                    userId: { [Op.notIn]: assignedUserId} 
+                }
+            })
+            const existing = await db.ProjectMember.findAll({
+                    where: { projectId: id }
+            });
+            const existingMemberIds = existing.map(row => row.userId);
+            console.log(existingMemberIds)
+            const newMembers = assignedUserId.filter(id => !existingMemberIds.includes(id));
+            console.log(newMembers)
+            await db.ProjectMember.bulkCreate(
+                newMembers.map(user_id => ({
+                    projectId: id,
+                    userId: user_id,
+                }))
+            );
+        } else{
+            throw new Error(`User does not have permission to edit this task.`)
+        }
     } catch (error){
         throw new Error(`check error ${error}`)
     }
@@ -147,13 +158,25 @@ const getProjectProgress=async(projectId)=>{
         throw new Error(`check error ${error}`)
     }
 }
-const deleteProject=async(projectId)=>{
+const deleteProject=async(userId, projectId)=>{
     try{
-       await db.Project.destroy({
-        where:{
-            id:projectId
+        const checkRole=await db.ProjectMember.findOne({
+            where:{
+                userId:userId,
+                projectId:projectId
+            }
+        })
+        const plainResult=await checkRole.get({plain:true})
+        if(plainResult.role==='admin'){
+            await db.Project.destroy({
+                where:{
+                    id:projectId
+                }
+            })
+        } else{
+            throw new Error(`User does not have permission to edit this task.`)
+
         }
-       })
     } catch (error){
         throw new Error(`check error ${error}`)
     }
